backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["sevenc-nanashi <sevenc7c@sevenc7c.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.88.0"
uvicorn = "^0.20.0"
gunicorn = "^20.1.0"
aioredis = "^2.0.1"
python-dotenv = "^0.21.0"
pycryptodome = "^3.16.0"
python-i18n = {extras = ["YAML"], version = "^0.3.9"}
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
